Move classes to `utils` package
