+++ src/ru/nsu/ccfit/zuev/osu/MainScene.java
@@ -9,8 +9,8 @@ import com.acivev.VibratorManager;
 import com.reco1l.andengine.Anchor;
 import com.reco1l.andengine.shape.Box;
 import com.reco1l.andengine.sprite.ExtendedSprite;
-import com.osudroid.BannerManager;
-import com.osudroid.BannerManager.BannerSprite;
+import com.osudroid.ui.BannerManager;
+import com.osudroid.ui.BannerManager.BannerSprite;
 import com.reco1l.osu.data.BeatmapInfo;
 import com.reco1l.osu.Execution;
 import com.osudroid.ui.MainMenu;
